requirements.txt
streamlit
openai
SpeechRecognition
pydub
python-dotenv